feat(engine): export is quality-driven — settings replace platform presets
RenderOptions.settings (resolution/fps/quality/codec/container) derives the
encode target from the project's aspect, capped at the sources; the encoder
takes the ladder bitrate; the job reports the target actually used. Named
platform presets, the Reels default and the frame-grab preset lookup are gone;
CLI takes --resolution/--fps/--quality/--codec/--container. Fixtures that
relied on the forced 9:16 default now state their aspect. 2706 tests green.

+++ engine/python/framepilot_engine/render/compiler.py
@@ -1051,7 +1051,7 @@ def compile_timeline(
 
 
 def _composite_with_blend_modes(
-    video_layers: list[tuple[Any, str | None]], target: tuple[int, int], fps: int
+    video_layers: list[tuple[Any, str | None]], target: tuple[int, int], fps: float
 ) -> VideoClip:
     from moviepy import CompositeVideoClip as _CompositeVideoClip
